refactor(proyect): first round of proyect refactor
